+++ src/Robotone/Processor.cpp
@@ -52,7 +52,7 @@ Processor::Processor()
     {
       if (effects[i])
       {
-        effects[i]->gestalt(std::get<double>(value));
+        effects[i]->gestalt(std::get<int>(value));
       }
     }
   });
@@ -65,7 +65,7 @@ std::unique_ptr<Effect> Processor::createEffect(const size_t channel, const doub
   effect->decimate(parameters.get<bool>("decimate"));
   effect->millis(parameters.get<int>("millis"));
   effect->octave(parameters.get<int>("octave"));
-  effect->gestalt(parameters.get<double>("gestalt"));
+  effect->gestalt(parameters.get<int>("gestalt"));
   effect->reset();
 
   return effect;
